Extract forcing versions
